refactor: 중복함수 분리
